fixed building _mysql.so with static link to libmysqlclient.a
libmysqlclient uses C++, so when linking against the static version of the library,
have to explicitly link against libstdc++. This is a dynamic link because statically linking
against libstdc / libstdc++ is generally a very bad idea, since they can have ties to the kernel.

added a flag to this build process to fail if there are undefined symbols while building.
with this flag, commenting out the addition of 'stdc++' as a library demonstrates the problem.