Always skip SSL if host=localhost.
If we are making a connection to localhost, then passing anything for
SSL will always mean a failed connection, because a certificate cannot
be verified (if TCP/IP) or SSL doesn't make sense (if UNIX).



Make it easier to call connect with ssl values by always ignoring them
in the localhost case. This means that callers do not need to check if
the host is localhost and can simply always provide an SSL value (e.g.
if the host is a variable).