Add support for converting sets
For a column defined as SET('A','B'), the C library returns
FIELD_TYPE_STRING and FLAG_SET instead of FIELD_TYPE_SET. This leads to
values being converted to a string (with comma-separated values) instead
of a proper Python set. Adding a mapping that watches for FLAG_SET fixes
the issue.

Signed-off-by: Stefanie Grunwald <steffi@physics.org>