memory leak on self->connection when mysql_real_connect fails
Looking at mysql_real_connect on failure it does not free the memory of the passed in MYSQL* struct. only mysql_close does that.
So we can't set open=0 after we mysql_init or dealloc will not cleanup the memory.

Also if mysql_init returns NULL we are out of memory and shouldn't set open=1, or we could segfault in dealloc if we didn't seg before that.