Use atexit to only shutdown MariaDB on module exit
The MariaDB embedded server can only be started once per process. You
can open as many connections as you want, but the starting of the server
can only happen once. Currently the start is protected fine, but on
connection close the server was shutdown. This meant that open and
closing connections was causing segfaults because MariaDB can't be
started again once it is stopped in embedded mode. The change here makes
the stopping of the server happen as global process level when the
python module is unloaded.